sort the files
